refactor(cli): extract PilotMcpServer + run loop into mcp/server.rs (issue #70)
Move struct + run_mcp_server + impl block + ServerHandler from _legacy
into mcp/server/{mod,methods}.rs (split because impl block alone exceeds
150 lines). call_tool_by_name extracted as a free function in
mcp/handlers/mod.rs taking &PilotMcpServer. _legacy.rs continues to
shrink toward zero (PR1 Tasks 6-8).
